rename AzureConfigurations and add supress action map property.
